refactored renderList into renderListWithTemplate
